backend/requirements-dev.txt
-r requirements.txt

# Testing
pytest
pytest-django
pytest-cov
pytest-mock
factory-boy
faker

# Typing / stubs
django-stubs
djangorestframework-stubs

# Code Quality
pyright
ruff
black
isort

# Documentation
sphinx
sphinx-rtd-theme
